# template/api/web/auth_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.32.0
// source: auth.proto

package web

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAuthGetLoginSignatureText = "/web.Auth/GetLoginSignatureText"
const OperationAuthLoginByWallet = "/web.Auth/LoginByWallet"

type AuthHTTPServer interface {
	// GetLoginSignatureText Get login signature text
	GetLoginSignatureText(context.Context, *GetLoginSignTextRequest) (*GetLoginSignTextResponse, error)
	// LoginByWallet Login by web3 wallet
	LoginByWallet(context.Context, *LoginByWalletRequest) (*LoginByWalletResponse, error)
}

func RegisterAuthHTTPServer(s *http.Server, srv AuthHTTPServer) {
	r := s.Route("/")
	r.POST("/auth/login/wallet", _Auth_LoginByWallet0_HTTP_Handler(srv))
	r.GET("/auth/login/sign_text", _Auth_GetLoginSignatureText0_HTTP_Handler(srv))
}

func _Auth_LoginByWallet0_HTTP_Handler(srv AuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginByWalletRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthLoginByWallet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LoginByWallet(ctx, req.(*LoginByWalletRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginByWalletResponse)
		return ctx.Result(200, reply)
	}
}

func _Auth_GetLoginSignatureText0_HTTP_Handler(srv AuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLoginSignTextRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthGetLoginSignatureText)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLoginSignatureText(ctx, req.(*GetLoginSignTextRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetLoginSignTextResponse)
		return ctx.Result(200, reply)
	}
}

type AuthHTTPClient interface {
	// GetLoginSignatureText Get login signature text
	GetLoginSignatureText(ctx context.Context, req *GetLoginSignTextRequest, opts ...http.CallOption) (rsp *GetLoginSignTextResponse, err error)
	// LoginByWallet Login by web3 wallet
	LoginByWallet(ctx context.Context, req *LoginByWalletRequest, opts ...http.CallOption) (rsp *LoginByWalletResponse, err error)
}

type AuthHTTPClientImpl struct {
	cc *http.Client
}

func NewAuthHTTPClient(client *http.Client) AuthHTTPClient {
	return &AuthHTTPClientImpl{client}
}

// GetLoginSignatureText Get login signature text
func (c *AuthHTTPClientImpl) GetLoginSignatureText(ctx context.Context, in *GetLoginSignTextRequest, opts ...http.CallOption) (*GetLoginSignTextResponse, error) {
	var out GetLoginSignTextResponse
	pattern := "/auth/login/sign_text"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAuthGetLoginSignatureText))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// LoginByWallet Login by web3 wallet
func (c *AuthHTTPClientImpl) LoginByWallet(ctx context.Context, in *LoginByWalletRequest, opts ...http.CallOption) (*LoginByWalletResponse, error) {
	var out LoginByWalletResponse
	pattern := "/auth/login/wallet"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthLoginByWallet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}