# template/api/web/probe.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.8
// 	protoc        v6.32.0
// source: probe.proto

package web

import (
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	structpb "google.golang.org/protobuf/types/known/structpb"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type ReadinessProbeResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Status        string                 `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ReadinessProbeResponse) Reset() {
	*x = ReadinessProbeResponse{}
	mi := &file_probe_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ReadinessProbeResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ReadinessProbeResponse) ProtoMessage() {}

func (x *ReadinessProbeResponse) ProtoReflect() protoreflect.Message {
	mi := &file_probe_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ReadinessProbeResponse.ProtoReflect.Descriptor instead.
func (*ReadinessProbeResponse) Descriptor() ([]byte, []int) {
	return file_probe_proto_rawDescGZIP(), []int{0}
}

func (x *ReadinessProbeResponse) GetStatus() string {
	if x != nil {
		return x.Status
	}
	return ""
}

var File_probe_proto protoreflect.FileDescriptor

const file_probe_proto_rawDesc = "" +
	"\n" +
	"\vprobe.proto\x12\x03web\x1a\x1cgoogle/api/annotations.proto\x1a\x1bgoogle/protobuf/empty.proto\x1a\x1cgoogle/protobuf/struct.proto\"0\n" +
	"\x16ReadinessProbeResponse\x12\x16\n" +
	"\x06status\x18\x01 \x01(\tR\x06status2\xab\x01\n" +
	"\x05Probe\x12K\n" +
	"\ahealthy\x12\x16.google.protobuf.Empty\x1a\x16.google.protobuf.Empty\"\x10\x82\xd3\xe4\x93\x02\n" +
	"\x12\b/healthy\x12U\n" +
	"\x05ready\x12\x17.google.protobuf.Struct\x1a\x1b.web.ReadinessProbeResponse\"\x16\x82\xd3\xe4\x93\x02\x10\x12\x0e/are/you/readyB1Z/github.com/carv-protocol/kratos-ddd/api/web;webb\x06proto3"

var (
	file_probe_proto_rawDescOnce sync.Once
	file_probe_proto_rawDescData []byte
)

func file_probe_proto_rawDescGZIP() []byte {
	file_probe_proto_rawDescOnce.Do(func() {
		file_probe_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_probe_proto_rawDesc), len(file_probe_proto_rawDesc)))
	})
	return file_probe_proto_rawDescData
}

var file_probe_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_probe_proto_goTypes = []any{
	(*ReadinessProbeResponse)(nil), // 0: web.ReadinessProbeResponse
	(*emptypb.Empty)(nil),          // 1: google.protobuf.Empty
	(*structpb.Struct)(nil),        // 2: google.protobuf.Struct
}
var file_probe_proto_depIdxs = []int32{
	1, // 0: web.Probe.healthy:input_type -> google.protobuf.Empty
	2, // 1: web.Probe.ready:input_type -> google.protobuf.Struct
	1, // 2: web.Probe.healthy:output_type -> google.protobuf.Empty
	0, // 3: web.Probe.ready:output_type -> web.ReadinessProbeResponse
	2, // [2:4] is the sub-list for method output_type
	0, // [0:2] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_probe_proto_init() }
func file_probe_proto_init() {
	if File_probe_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_probe_proto_rawDesc), len(file_probe_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_probe_proto_goTypes,
		DependencyIndexes: file_probe_proto_depIdxs,
		MessageInfos:      file_probe_proto_msgTypes,
	}.Build()
	File_probe_proto = out.File
	file_probe_proto_goTypes = nil
	file_probe_proto_depIdxs = nil
}